+++ cpu/includes/initial_setup.h
@@ -23,7 +23,7 @@ extern char* memoria_port;
 
 extern char* server_port;
 
-extern int retardo_instruc;
+extern double retardo_instruc;
 
 extern int tam_max_segmento;
